Tests moved to gamearea package so PGGameArea object hand could stay protected
